Add covid and pool_testing scriptlets.
This is by no means ready for widespread consumption, as there's a
number of hardcoded constants that would be better suited as .

covid.py uses git submodules for Johns Hopkins and NYT data, as well
as COVID Tracking Project data fetched from its web API, to graph new
daily cases on a per-state basis in the US.

It also provides a manually-curated breakdown of states that had a >5%
margin in the 2016 election, and an autogenerated set of counties that
had a >10% margin (editable in populate_Counties()), to see if there
was a significant partisan difference in areas that skewed heavily in
one direction or the other. Said differences existed early, as
strongly Democratic urban areas such as Chicago, NYC, the SF Bay Area,
and Boston were hit early on, but since then, there hasn't been a
significant difference.

Also note that regardless of threshold, the strongly partisan
population for each of the major parties seems to be relatively close,
both on a per-state and per-county basis.

pool_testing provides a number of functions for evaluating the
effectiveness / optimality thresholds for group testing, using a naive
2-shot group testing strategy.

Note that the modeling assumes that test results are iid Bernoulli
(distributed across individuals in a population, who are thus likely
to have their samples mixed).

More sophisticated group testing strategies may stretch out the
effectiveness of group testing, e.g. increasing the test coverage by
using each sample more than once.